Move RVprec with clumping to legacy.py.
