+++ eniric_scripts/aces_precision.py
@@ -12,7 +12,8 @@ import eniric
 from eniric.atmosphere import Atmosphere
 from eniric.broaden import convolution
 from eniric.corrections import correct_artigau_2018
-from eniric.Qcalculator import RVprec_calc, RVprec_calc_masked, quality
+from eniric.legacy import RVprec_calc_masked
+from eniric.Qcalculator import RVprec_calc, quality
 from eniric.resample import log_resample
 from eniric.snr_normalization import snr_constant_band
 from eniric.utilities import band_middle, load_aces_spectrum, load_btsettl_spectrum
@@ -175,7 +176,7 @@ def do_analysis(
         )
 
     wav_grid, sampled_flux = convolve_and_resample(
-        wav, flux, vsini, R, band, sampling, conv_kwargs
+        wav, flux, vsini, R, band, sampling, **conv_kwargs
     )
 
     # Spectral Quality
@@ -183,7 +184,7 @@ def do_analysis(
 
     # Scale normalization for precision
     wav_ref, sampled_ref = convolve_and_resample(
-        wav, flux, vsini, R, ref_band, sampling, conv_kwargs
+        wav, flux, vsini, R, ref_band, sampling, **conv_kwargs
     )
     snr_normalize = snr_constant_band(wav_ref, sampled_ref, snr=snr, band=ref_band)
     sampled_flux = sampled_flux / snr_normalize
@@ -224,7 +225,7 @@ def convolve_and_resample(
     R: float,
     band: str,
     sampling: float,
-    conv_kwargs,
+    **conv_kwargs,
 ) -> Tuple[ndarray, ndarray]:
     """Convolve and resample functions together.
